requirements.txt
PyYAML>=6.0
Jinja2>=3.1
oracledb>=2.0.1
